% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtum.R
\name{gtum}
\alias{gtum}
\title{Generalized Thurstonian Unfolding Model estimation}
\usage{
gtum(
  gtum.Data,
  ind,
  ParInits,
  block,
  pairmap = NULL,
  covariate = NULL,
  iter = 2000,
  chains = 2,
  warmup = floor(iter/2),
  adapt_delta = 0.85,
  thin = 1,
  cores = 2,
  ma = 0.2,
  va = 1,
  mt = 0,
  vt = 3,
  mdne = -2,
  vdne = 2,
  mdnu = 0,
  vdnu = 2,
  mdpo = 2,
  vdpo = 2,
  mb = 0,
  vb = 1
)
}
\arguments{
\item{gtum.Data}{Response data in wide format (each row represents a person). If the original block size is three or more, then users need to first decompose the original responses into several pairwise comparisons and use the pairwise comparison data as input here. For example, if the original test has 3 blocks and each block contains statements A, B, and C, then the input data should have 9 columns (A1B1, A1C1, B1C1, A2B2, A2C2, B2C2, A3B3, A3C3, B3C3) where 1 means the first statement within the pair is preferred and 2 means the second statement within the pair is preferred. For graded preference forced-choice design, the data should be coded as the response option endorsed (e.g., 1= Statement A is much more like me; 2= Statement A is slightly more like me; 3= Statement B is slightly more like me; 4= Statement B is much more like me).}

\item{ind}{A two-column matrix mapping each statement to each trait in the pair format. For example, matrix(c(1, 1, 1, 2, 2, 2), ncol = 2) means that for each pair, the first statement measures trait 1 and the second statement measures trait 2.}

\item{ParInits}{A three-column matrix containing initial values for the statement parameters block by block. 1 and -1 for alphas and taus are recommended and -1/-2 or 1/2 for deltas are recommended depending on the signs of the statements. Pre-estimated statement parameters can be used as the initial values for scoring purpose.}

\item{block}{The number of statements in each block in the original test. For now, it can be 2 or 3. We will further expand the code to incorporate bigger blocks.}

\item{pairmap}{A two-column matrix specifying the ID of statements within each trait. The row of this matrix equals to the total number of pairwise comparisons. For example, for a 20-block tests with 3 statements per block, there will be 20*(3*(3-1)/2) = 60 pairwise comparisons. Suppose there are 12 statements measuring each trait. Then 1 means the statement is the first statement measuring the trait and 12 means the statement is the twelfth statement measuring the trait.}

\item{covariate}{An p*c person covariate matrix where p equals sample size and c equals the number of covariates. The default is NULL, meaning no person covariate.}

\item{iter}{The number of iterations. The default value is 2000. See documentation for rstan for more details.}

\item{chains}{The number of chains. The default value is 2. See documentation for rstan for more details.}

\item{warmup}{The number of warmups to discard. The default value is 0.5*iterations. See documentation for rstan for more details.}

\item{adapt_delta}{Target average proposal acceptance probability during Stan's adaptation period. The default value is 0.85. See documentation for rstan for more details.}

\item{thin}{Thinning. The default value is 1. See documentation for rstan for more details.}

\item{cores}{The number of computer cores used for parallel computing. The default value is 2.}

\item{ma}{Mean of the prior distribution for alphas, which follows a lognormal distribution. The default value is 0.2.}

\item{va}{Standard deviation of the prior distribution for alpha. The default value is 1.}

\item{mt}{Means of the prior distributions for taus, which follows a normal distribution. The default values is 0.}

\item{vt}{Standard deviation of the prior distribution for taus. The default value is 3.}

\item{mdne}{Mean of the prior distribution for the location parameters (delta) of negative items. For example, the item “I am often lazy” as an indicator of conscientiousness should have a negative location parameter (e.g., -2.5). Location parameters are on the same scale (z score) as the latent trait scores. We assume a normal prior for the location parameters and fix the range of negative location parameters between -10 and 0. The default value is -2.}

\item{vdne}{Standard deviation of the prior distribution for negative deltas. The default value is 2.}

\item{mdnu}{Mean of the prior distribution for the location parameters (delta) of neutral (intermediate) items. For example, the item “I as productive as an average person” as an indicator of conscientiousness should have a location parameter close to zero. We fix the range of location parameters of neutral items between -10 and 10 as we are not sure whether they are slightly positive or slightly negative. The default value is 0.}

\item{vdnu}{Standard deviation of the prior distribution for negative deltas. The default value is 2.}

\item{mdpo}{Mean of the prior distribution for the location parameters (delta) of positive items. For example, the item “I very hardworking” as an indicator of conscientiousness should have a positive location parameter (e.g., 2.5). We fix the range of location parameters of positive items between 0 and 10.  The default value is 2.}

\item{vdpo}{Standard deviation of the prior distribution for positive deltas. The default value is 2.}

\item{mb}{Means of the prior distributions for the random block factor. The default values is 0.}

\item{vb}{Standard deviation of the prior distribution for the random block factor. The default value is 1.}
}
\value{
Result object that stores information including the (1) stanfit object, (2) estimated item parameters, (3) estimated person parameters, (4) response data, and (5) the input column vector mapping each statement to each trait.
}
\description{
This function implements full Bayesian estimation of the Generalized Thurstonian Unfolding Model using rstan
}
\examples{
\donttest{
Data <- c(1,5)
Data <- matrix(Data,nrow = 1)
ind <- matrix(c(1, 1, 2, 3), ncol = 2)
ParInits <- matrix(c(-1, -1, -1, -1), ncol = 2)
mod <- gtum(gtum.Data=Data,ind=ind,ParInits=ParInits,block=2,iter=3,warmup=1,chains=1)}
}
