% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtum.R
\name{gtum}
\alias{gtum}
\title{Generalized Thurstonian Unfolding Model estimation}
\usage{
gtum(
  gtum.Data,
  ind,
  ParInits,
  block,
  pairmap = NULL,
  covariate = NULL,
  iter = 2000,
  chains = 2,
  warmup = floor(iter/2),
  adapt_delta = 0.85,
  thin = 1,
  cores = 2,
  ma = 0.2,
  va = 0.5,
  mt = 0,
  vt = 2,
  ma_t = 0.2,
  va_t = 1,
  mt_t = 0,
  vt_t = 3,
  mdne = 0.7,
  vdne = 0.2,
  mdnu = 0.5,
  vdnu = 0.1,
  mdpo = 0.3,
  vdpo = 0.2
)
}
\arguments{
\item{gtum.Data}{Response data in wide format. For dichotomous forced choice data, if the first statement is preferred, the data should be coded as 1, otherwise it should be coded as 2. For polytomous forced choice data, the data should be coded as the response option endorsed.}

\item{ind}{A two-column matrix mapping each statement to each trait. For example, matrix(c(1, 1, 1, 2, 2, 2), ncol = 2) means that for each pair, the first statement measures trait 1 and the second statement measures trait 2.}

\item{ParInits}{A two-column or three column (depending on the block size) matrix specifying the directions or positivity/negativity of the statements in each pair. 1 means the statement is positive, -1 means the statement is negative, and 0 means the statement is neutral(intermediate items). For example, matrix(c(1, 1, 1, -1, -1, -1), ncol = 2) means that for each pair, the first statement is positive and the second statement is negative.}

\item{block}{The number of statements in each block in the original test. For now, it can be 2 or 3. We will further expand the code to incorporate bigger blocks.}

\item{pairmap}{A two-column matrix specifying the rank/ID of the statement in each trait it measures. For example, suppose there are 3 statements measuring each trait. 1 means the statement is the first statement measuring the trait and 3 means the statement is the last statement measuring the trait.}

\item{covariate}{An p*c person covariate matrix where p equals sample size and c equals the number of covariates. The default is NULL, meaning no person covariate.}

\item{iter}{The number of iterations. The default value is 2000. See documentation for rstan for more details.}

\item{chains}{The number of chains. The default value is 2. See documentation for rstan for more details.}

\item{warmup}{The number of warmups to discard. The default value is 0.5*iterations. See documentation for rstan for more details.}

\item{adapt_delta}{Target average proposal acceptance probability during Stan's adaptation period. The default value is 0.85. See documentation for rstan for more details.}

\item{thin}{Thinning. The default value is 1. See documentation for rstan for more details.}

\item{cores}{The number of computer cores used for parallel computing. The default value is 2.}

\item{ma}{Mean of the prior distribution for alphas for block of 2 in the original test, which follows a lognormal distribution. The default value is 0.2.}

\item{va}{Standard deviation of the prior distribution for alpha for block of 2 in the original test. The default value is 0.5.}

\item{mt}{Means of the prior distributions for taus for block of 2 in the original test, which follows a normal distribution. The default values is 0.}

\item{vt}{Standard deviation of the prior distribution for taus for block of 2 in the original test. The default value is 2.}

\item{ma_t}{Mean of the prior distribution for alphas for block of 3 in the original test, which follows a lognormal distribution. The default value is 0.2.}

\item{va_t}{Standard deviation of the prior distribution for alpha for block of 3 in the original test. The default value is 1.}

\item{mt_t}{Means of the prior distributions for taus for block of 3 in the original test, which follows a normal distribution. The default values is 0.}

\item{vt_t}{Standard deviation of the prior distribution for taus for block of 3 in the original test. The default value is 3.}

\item{mdne}{Mean of the prior distribution for negative deltas for block of 2/3 in the original test, which follows a normal distribution. The default value is 0.7.}

\item{vdne}{Standard deviation of the prior distribution for negative deltas for block of 2/3 in the original test. The default value is 0.2.}

\item{mdnu}{Mean of the prior distribution for neutral deltas for block of 2/3 in the original test, which follows a normal distribution. The default value is 0.5.}

\item{vdnu}{Standard deviation of the prior distribution for neutral deltas for block of 2/3 in the original test. The default value is 0.1.}

\item{mdpo}{Mean of the prior distribution for positive deltas for block of 2/3 in the original test, which follows a normal distribution. The default value is 0.3.}

\item{vdpo}{Standard deviation of the prior distribution for positive deltas for block of 2/3 in the original test. The default value is 0.2.}
}
\value{
Result object that stores information including the (1) stanfit object, (2) estimated item parameters, (3) estimated person parameters, (4) response data, and (5) the input column vector mapping each statement to each trait.
}
\description{
This function implements full Bayesian estimation of the Generalized Thurstonian Unfolding Model using rstan
}
\examples{
\donttest{
Data <- c(1,5)
Data <- matrix(Data,nrow = 1)
ind <- matrix(c(1, 1, 2, 3), ncol = 2)
ParInits <- matrix(c(-1, -1, -1, -1), ncol = 2)
mod <- gtum(gtum.Data=Data,ind=ind,ParInits=ParInits,block=2,iter=3,warmup=1,chains=1)}
}
